feat(readme): Add references to the README
I'm adding a list of all references so that others can know where these templates come from. I know the `feat` tag is usually not used for commits like this but I would like to trigger a new release and build of the docs to make sure that everything is working.