feat(data): Simplify package to use Honeybee JSON data
This commit is coordinated with recent changes to honeybee_energy that mean the data folder is the only thing needed to actually have this function as an extension of the standards. As such, all of the Python code of this repo has been reformatted so that it only gets loaded whenever we need to update this data with new additions to the OpenStudio standards gem or because of changes in the Honeybee schema.