fix(dependabot): Use >= operator for honeybee-energy version
It is very rare that we need a new release of this repo. Only specific edits to the schema of materials, constructions, constructionsets, schedules, or program types should merit a new release.

So I am switching this repo to use >= operators for it's honeybee-energy dependency and telling dependabot to not send a new PR whenever there is a release of honeybee-energy.