fix(deps): Move honeybee-energy to dev-requirements
Otherwise, we will get all of the dependency tree down to ladybug-geometry when we install in the resources/standards folder. I think we will just have to say that this repo only uses honeybee-energy for tests and, because ti's only data now, it doesn't require any version of honeybee-energy to run.