fix(update): Re-export all standards and improve ASHRAE-2004
This commit includes a re-export of all the standards. It also adds the baseline_watts_per_area property to each of the Lighting objects, which will be useful in baseline model creation workflows.

Lastly, it changes the functions that produce ConstructionSets for ASHRAE-2004 to use the constructions associated with 40% window ratios.  This should be a more realistic starting point for users than the 10% window ratio constructions used previously.