feat(ashrae): Add programs and construction sets for 2016 and 2019
Now that the latest versions of ASHRAE 90.1 were added to openstudio-standards, this commit adds these new versions to this package. I am also streamlining the process of adding the soil materials to the library and putting in some code in preparation for the eventual integration of the College Building Type.